absorb Nez into this repo instead of as a submodule
this makes a variety of things simpler when using vscode, since it seems to love restoring stuff it shouldn't be restoring, and adding a bunch of stuff dotnet gets confused over. This has the effect of breaking the build every time I use vscode to look at one of the Nez files, which is something I do often.